requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
openai==2.21.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pypdf==5.4.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
